+++ skimage/segmentation/_join.py
@@ -1,6 +1,7 @@
 import numpy as np
 from .._shared.utils import deprecated
 from ..util import map_array
+from ..util._map_array import ArrayMap
 
 
 def join_segmentations(s1, s2):
