+++ skimage/segmentation/setup.py
@@ -15,7 +15,7 @@ def configuration(parent_package='', top_path=None):
             '_felzenszwalb_cy.pyx',
             '_quickshift_cy.pyx',
             '_slic.pyx',
-            '_remap.pyx'], working_path=base_path)
+            ], working_path=base_path)
     config.add_extension('_watershed_cy', sources=['_watershed_cy.c'],
                          include_dirs=[get_numpy_include_dirs()])
     config.add_extension('_felzenszwalb_cy', sources=['_felzenszwalb_cy.c'],
@@ -24,12 +24,6 @@ def configuration(parent_package='', top_path=None):
                          include_dirs=[get_numpy_include_dirs()])
     config.add_extension('_slic', sources=['_slic.c'],
                          include_dirs=[get_numpy_include_dirs()])
-    # note: the extra compiler flag -std=c++0x is needed to access the
-    # std::unordered_map container on some earlier gcc compilers. See:
-    # https://stackoverflow.com/a/3973692/224254
-    config.add_extension('_remap', sources='_remap.cpp',
-                         include_dirs=[get_numpy_include_dirs()],
-                         language='c++', extra_compile_args=['-std=c++0x'])
 
     return config
